Remove useless interface PillarboxExoPlayer
